feat (bstimepicker / bsdatepicker): add change allowing support for typing / return viewValue for custom validation
This feature allows developers to specify their own validation by essentially disabling the internal validation. It still passes through an instanceof Date when the bstimepicker / bsdatepicker control is used, but allows $watch to work in consuming controllers so one can offer manual validation of dates / times
This change does not affect the normal operation of the bsdatepicker directive.